Redo generated type structure to better support type inference.
This makes the types generated "Complete", allowing for much better
inference. but makes it harder to extend with external values.
We'll think about this later.

At a minimum, we need to extend the schema we import with a list of
custom triples with common enums that are external to schema.org
altogether.