Fix types derived transitively from DataType.
A class that inherits from DataType:
1. Is not represented by a JSON-LD node type
2. Does not have any properties on it
3. Does not have a "@type"-tagged leaf type.

Take URL for example, which inerits from Text, a DataType
representing a string. URL can only be represented as a string.

The Class.prototype.aliasesBuiltin method was unused, therein
lied the problem. Use that method to decide if the nonEnumType
should contain a leaf type or not.

There's an overdue refactor here to clean up these types.