Fold most uses of gulp into simple npm scripts.
There's a bunch of gulp logic that's realy just straightforward in bash.
So we move it to npm scripts. Removing the need to have gulp installed
globally, and removing the need for gulp-typescript, which had its host
of issues and incompatibilities with tsc --b and some of the new
package updates we introduced.