Remove unrechable code from toClass.
We check for classes.size right after adding wellKnownTypes to that set.
It will never be 0.

The assertion in BuildClasses will never be true, since IsClass is the
same condition as what causes this content to be added. Instead of
bespoke exceptions, we use node's assertion to make sure that this won't
happen.

More philosophically, we should use this type of assertion whenever --
within the bounds of an exported function -- regardless of its inputs,
a certain condition is impossible.

Instead let's just make sure explicit exceptions are used for things that
can go wrong in actuality.