Reader Coverage: Restructure assertion code.
Per the standard established in #73 (specifically in 11c9f643), use an
external assert function using TypeScript's "asserts" returns for areas
where we structurally expect something to never happen.

In this code, the regex used structurally makes it so that verify will
never throw:

- The first two unWrap() parsing options will only work if the thing
  is wrapped in <>. If it is, but doesn't work, UrlNode throws.
- The the last function throws on malformed string, but the regex
  guarantees that the string is well-formed.

Change this into an assertion. If it happens, its because a structural
assumption made about the code is wrong.

As opposed to other exception places where the error happens not because
of bugs in the code, but because of issues with the data or the inputs.

This factors out assert into its own util file.

Verify is only used in object(), so we inline its logic.