Restructure comment handling code.
There was an impossibility in error handling where the onCloseTag
branch's default (throw Error) statement was impossible. This is because
we would always throw in onOpenTag first. We will never see a valid
closed tag without an opening.

Further, a lone closed tag will fail to parse as a close tag anyway. So
this just never happened.

One might be tempted to just remove the default in the name of coverage.
But the default line actually gives us some value: If we ever add
something in open, we might know if we forgot ot add it in close.

But that seems like it should be enforced on lint-time or compile-time,
not just if we're lucky to exercise a condition. Ideally, we should let
the compiler guarantee that any opening tag has a corresponding closing
tag. This way, we keep the benefits of this line, and even improve them
to become a compile-time property, while changing the closing logic
to reflect the impossibility of seeing an unknown closing tag.