Baseline: Run entire main function for baseilne tests.
This way, we no longer have our own implementation of getTriples or
directly calling WriteDeclarations and others.

There's a fair amount of fake-stubs and mocks to get this working (right
now, our source can only URLs with an HTTPS protocol), though over time
with native local-file support, the hope is we'd have to do much less of
that.