Allow `{"@id": url}` references.
For any property position in which the value can be a node type, allow
the use of a standalone "@id" reference in its place.

A plain "@id" object is in many ways the bottom type of the JSON-LD node
hierarcy. (Though not the SDO DataType hierarcy, nulls, etc)

BREAKING: Property parsing is stricter on custom schema where
rangeIncludes references a non-existent type.