Make each package control its own linting.
For now, we keep eslintrc at the very top. In the future we might do
something different (with a shared rule set at the root), maybe.

We might want to go the opposite approach: control linting always at the
top level. This is what you'd have in a large monorepo with a consistent
style guide. But eslint (especially with typescript-eslint), is brittle
when doing that; we want to lint **source** files, but not **generated**
files. ESLint also needs to be passed exactly what file(s) to operate
on. This means that we end up in an annoying place:

- ESLint needs to know what all of your source files what (vs generated
  files).
- The script calling ESLint needs to know what all of your source files
  are.

Your guess is as good as mine for why one of these isn't sufficient.
Maybe it is, but I'm doing this wrong? Or typescript-eslint makes things
harder?

But anyway, here we are. We are now in a place where a global config
needs to know about the layout of each package. This seems difficult, so
we instead delegate this task to individual packages, who actually know
their own layout.