Refactor: natively operate on RDF graph
schema-dts-gen doesn't need it's own understanding of triples, and we
can simply use n3.js to traverse our RDF graph and get the
details we need.

This is a backwards-compatible change for the generator CLI itself, but
not any user of schema-dts-gen that directly imports its classes and
manipulates them.

The output is only subtly different in a few places, where the sorting
of parent classes is improved.