Remove TypeScript as an explicit peer dependency.
This is one potential fix to #180 that allows schema-dts to continue to be used
with `import` and any/no bundlers without increasing bundle size.

It is also possible to address #180 by simply asking users to save
schema-dts as a dev dependency. This is generally more correct for
typings, but runs the risk of leaving dangling imports to nonexistent
"schema-dts" js.

We _may_ want to add TypeScript as a dev dependency instead. I'm not
sure.